Add condition to deploy/1.dockerBuildEc2.sh file

+++ devops/build/1.dockerBuildEc2.sh
@@ -1,2 +1,4 @@
+# docker build
 docker build -t member:latest .
-docker rmi $(docker images -f dangling=true -q)
+# remove dangling image
+if [ $(docker images -f dangling=true -q | wc -l) -gt 1 ]; then docker rmi $(docker images -f dangling=true -q); fi
